Flowing water, volumetric water

+++ engine/include/engine/core/rendering/iWaterSurfaceSource.h
@@ -7,14 +7,16 @@ namespace sfs
 {
 
 // One column of water to draw a surface for: the tile, the water surface level
-// (elevation levels), and the solid floor level beneath it. Depth = surface -
-// floor. Render-style-agnostic, so both the entity-based water (WaterTile
-// component) and a voxel world can feed the same water surface mesh.
+// and the solid floor level beneath it, both in elevation levels and FRACTIONAL
+// (a per-cell water amount can rest at any height, e.g. on a half block).
+// Depth = surface - floor. Render-style-agnostic, so both the entity-based
+// water (WaterTile component) and a voxel world can feed the same water surface
+// mesh.
 struct WaterColumn
 {
   glm::ivec2 tile{0, 0};
-  int surfaceLevel = 0;
-  int floorLevel = 0;
+  float surfaceLevel = 0.0f;
+  float floorLevel = 0.0f;
 };
 
 // Supplies the water columns to render this frame. A voxel world implements it
